Modify tests

+++ src/test/unit/alignsensor_unittest.cc
@@ -40,9 +40,9 @@ extern "C" {
 
 #define DEG2RAD 0.01745329251
 
-static void rotateVector(int32_t mat[3][3], int32_t vec[3], int32_t *out)
+static void rotateVector(float mat[3][3], float vec[3], float *out)
 {
-    int32_t tmp[3];
+    float tmp[3];
 
     for(int i=0; i<3; i++) {
         tmp[i] = 0;
@@ -70,7 +70,7 @@ static void rotateVector(int32_t mat[3][3], int32_t vec[3], int32_t *out)
 //    mat[2][2] =  cos(angle*DEG2RAD);
 //}
 
-static void initYAxisRotation(int32_t mat[][3], int32_t angle)
+static void initYAxisRotation(float mat[][3], int32_t angle)
 {
     mat[0][0] =  cos(angle*DEG2RAD);
     mat[0][1] =  0;
@@ -83,7 +83,7 @@ static void initYAxisRotation(int32_t mat[][3], int32_t angle)
     mat[2][2] =  cos(angle*DEG2RAD);
 }
 
-static void initZAxisRotation(int32_t mat[][3], int32_t angle)
+static void initZAxisRotation(float mat[][3], int32_t angle)
 {
     mat[0][0] =  cos(angle*DEG2RAD);
     mat[0][1] = -sin(angle*DEG2RAD);
@@ -155,15 +155,15 @@ static void testCW(sensor_align_e rotation, int32_t angle)
  */
 static void testCWFlip(sensor_align_e rotation, int32_t angle)
 {
-    int32_t src[XYZ_AXIS_COUNT];
-    int32_t test[XYZ_AXIS_COUNT];
+    float src[XYZ_AXIS_COUNT];
+    float test[XYZ_AXIS_COUNT];
 
     // unit vector along x-axis
     src[X] = 1;
     src[Y] = 0;
     src[Z] = 0;
     
-    int32_t matrix[3][3];
+    float matrix[3][3];
     initYAxisRotation(matrix, 180);
     rotateVector(matrix, src, test);
     initZAxisRotation(matrix, angle);
